requirements.txt
Django>=4.2
djangorestframework
graphene-django
celery
bleach
libsass
colormath
requests
